requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-login==0.6.2
flask-wtf==1.1.1
flask-sqlalchemy==3.0.5
email-validator==2.0.0
python-dotenv==1.0.0
